refactor(repl): migrate macro-flow commands into MacroFlowCommandSet
Move gen_tile_macro, gen_all_tile_macros, gen_fabric_macro and
run_FABulous_eFPGA_macro into MacroFlowCommandSet, along with the
_require_directional_mode / _resolve_directional_fix helpers they use.
Update the macro tests to patch is_pdk_config_set and import
_resolve_directional_fix from their new home in cmd_macro.

stack-info: PR: https://github.com/FPGA-Research/FABulous/pull/892, branch: KelvinChung2000/stack/6
